Test that plugin works without optional dependency
This moves all tests testing the saml integration into a new class and
runs all other tests with & without this dependency.
